chore(databases-collections-list, compass-components): Extract virtual grid to reusable component in compass-components COMPASS-5408 (#2693)
* chore(databases-collections-list, compass-components): Extract virtual grid to reusable component in compass-components

* fix(databases-collections-list): Return testid to the grid container

* chore(compass-components): Remove duplicated mergeProps helper

* test(e2e): Fix database card selector

* chore(compass-components): Improve types

* chore(eslint-config-compass): Disable required propTypes for components written in typescript
