+++ packages/compass-components/package.json
@@ -67,7 +67,8 @@
     "@react-aria/interactions": "^3.6.0",
     "@react-aria/tooltip": "^3.1.3",
     "@react-aria/utils": "^3.9.0",
-    "@react-stately/tooltip": "^3.0.5"
+    "@react-stately/tooltip": "^3.0.5",
+    "react-window": "^1.8.6"
   },
   "peerDependencies": {
     "react": "^16.14.0"
