+++ packages/databases-collections-list/package.json
@@ -53,8 +53,7 @@
     "@mongodb-js/compass-components": "^0.9.1",
     "@react-aria/utils": "^3.10.0",
     "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-window": "^1.8.6"
+    "react-dom": "^16.14.0"
   },
   "devDependencies": {
     "@mongodb-js/compass-logging": "^0.6.1",
