+++ packages/databases-collections-list/src/collections.tsx
@@ -10,16 +10,6 @@ type Collection = {
   name: string;
   type: string;
   status: 'initial' | 'fetching' | 'refreshing' | 'ready' | 'error';
-  statusError: string | null;
-  readonly: boolean;
-  view_on: string;
-  collation: unknown;
-  pipeline: unknown[];
-  validation: unknown;
-  is_capped: boolean;
-  max: number;
-  is_power_of_two: boolean;
-  index_sizes: unknown;
   document_count: number;
   document_size: number;
   avg_document_size: number;
@@ -27,14 +17,7 @@ type Collection = {
   free_storage_size: number;
   index_count: number;
   index_size: number;
-  padding_factor: number;
-  extent_count: number;
-  extent_last_size: number;
-  flags_user: number;
-  max_document_size: number;
   size: number;
-  index_details: unknown;
-  wired_tiger: unknown;
   properties: { name: string }[];
   source?: Collection;
 };
