+++ packages/databases-collections-list/src/items-grid.tsx
@@ -1,23 +1,11 @@
 /* eslint-disable react/prop-types */
-import React, {
-  useMemo,
-  forwardRef,
-  useContext,
-  useRef,
-  useCallback,
-} from 'react';
-import { FixedSizeList } from 'react-window';
-import { css, cx, spacing, useDOMRect } from '@mongodb-js/compass-components';
+import React, { useContext, useCallback } from 'react';
+import { css, cx, spacing, VirtualGrid } from '@mongodb-js/compass-components';
 import { createLoggerAndTelemetry } from '@mongodb-js/compass-logging';
 import { useSortControls, useSortedItems } from './use-sort';
 import type { NamespaceItemCardProps } from './namespace-card';
 import { useViewTypeControls, ViewType } from './use-view-type';
 import { useCreateControls } from './use-create';
-import { mergeProps } from './merge-props';
-import {
-  useVirtualGridArrowNavigation,
-  useVirtualRovingTabIndex,
-} from './use-virtual-grid';
 
 const { track } = createLoggerAndTelemetry(
   'COMPASS-DATABASES-COLLECTIONS-LIST-UI'
@@ -29,78 +17,8 @@ const row = css({
   paddingLeft: spacing[3],
   paddingRight: spacing[3],
   paddingBottom: spacing[2],
-  display: 'grid',
-  alignItems: 'stretch',
-  gridTemplateRows: 'auto',
-  gridTemplateColumns: '1fr',
-  gridAutoColumns: '1fr',
-  gridAutoFlow: 'column',
   columnGap: spacing[2],
 });
-
-const cell = css({
-  width: '100%',
-  minWidth: 0,
-});
-
-type ItemRowProps<T extends Item> = {
-  style: React.CSSProperties;
-  index: number;
-  data: CallbackProps & {
-    items: T[];
-    currentTabbable: number;
-    viewType: ViewType;
-    rowsCount: number;
-    columnsCount: number;
-    renderItem: RenderItem<T>;
-  };
-};
-
-const ItemRow = <T extends Item>({
-  style,
-  index,
-  data,
-}: ItemRowProps<T>): React.ReactElement => {
-  // eslint-disable-next-line @typescript-eslint/unbound-method
-  const {
-    items,
-    columnsCount,
-    currentTabbable,
-    onItemClick,
-    onDeleteItemClick,
-    renderItem,
-    viewType,
-  } = data;
-  const rowStart = index * columnsCount;
-  const cells = items.slice(rowStart, rowStart + columnsCount);
-  const placeholderCellsCount = columnsCount - cells.length;
-  const emptyCells = useMemo(() => {
-    return Array.from({ length: placeholderCellsCount }, (_, idx) => (
-      <div className={cell} key={`empty${idx}`}></div>
-    ));
-  }, [placeholderCellsCount]);
-  return (
-    <div style={style} className={row} role="row" aria-rowindex={index + 1}>
-      {cells.map((item, index) => {
-        const idx = rowStart + index;
-        return renderItem({
-          item,
-          onItemClick,
-          onDeleteItemClick,
-          viewType,
-          className: cell,
-          tabIndex: idx === currentTabbable ? 0 : -1,
-          role: 'gridcell',
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-expect-error data-* attributes are unrecognised by HTMLProps
-          // type
-          ['data-vlist-item-idx']: idx,
-        });
-      })}
-      {emptyCells}
-    </div>
-  );
-};
 
 const container = css({
   width: '100%',
@@ -124,10 +42,6 @@ const controls = css({
 
 export const createButton = css({
   whiteSpace: 'nowrap',
-});
-
-const listContainer = css({
-  position: 'relative',
 });
 
 const control = css({
@@ -182,41 +96,24 @@ const ControlsContext = React.createContext<{
   createControls: React.ReactElement | null;
   viewTypeControls: React.ReactElement | null;
   sortControls: React.ReactElement | null;
-  gridProps: React.HTMLProps<HTMLDivElement>;
 }>({
   createControls: null,
   viewTypeControls: null,
   sortControls: null,
-  gridProps: {},
 });
 
-const ListWithControls = forwardRef<HTMLDivElement>(function ListWithControls(
-  { style, children, ...props }: React.HTMLProps<HTMLDivElement>,
-  ref
-) {
-  const { createControls, viewTypeControls, sortControls, gridProps } =
+const GridControls = () => {
+  const { createControls, viewTypeControls, sortControls } =
     useContext(ControlsContext);
 
   return (
-    <div
-      ref={ref}
-      style={{
-        ...style,
-        height: parseFloat(style?.height as string) + CONTROLS_HEIGHT,
-      }}
-      {...props}
-    >
-      <div className={controls}>
-        {createControls && <div className={control}>{createControls}</div>}
-        <div className={control}>{viewTypeControls}</div>
-        <div className={cx(control, pushRight)}>{sortControls}</div>
-      </div>
-      <div className={listContainer} {...gridProps}>
-        {children}
-      </div>
-    </div>
+    <>
+      {createControls && <div className={control}>{createControls}</div>}
+      <div className={control}>{viewTypeControls}</div>
+      <div className={cx(control, pushRight)}>{sortControls}</div>
+    </>
   );
-});
+};
 
 export const ItemsGrid = <T extends Item>({
   itemType,
@@ -229,7 +126,7 @@ export const ItemsGrid = <T extends Item>({
   onItemClick,
   onDeleteItemClick,
   onCreateItemClick,
-  renderItem,
+  renderItem: _renderItem,
 }: ItemsGridProps<T>): React.ReactElement => {
   const onViewTypeChange = useCallback(
     (newType) => {
@@ -237,42 +134,32 @@ export const ItemsGrid = <T extends Item>({
     },
     [itemType]
   );
-  const listRef = useRef<FixedSizeList | null>(null);
   const createControls = useCreateControls(itemType, onCreateItemClick);
   const [sortControls, sortState] = useSortControls(sortBy);
   const [viewTypeControls, viewType] = useViewTypeControls({
     onChange: onViewTypeChange,
   });
   const sortedItems = useSortedItems(items, sortState);
-  const [rectProps, { width, height }] = useDOMRect();
 
   const itemWidth = viewType === 'grid' ? itemGridWidth : itemListWidth;
   const itemHeight = viewType === 'grid' ? itemGridHeight : itemListHeight;
-  const colCount =
-    viewType === 'grid' ? Math.max(1, Math.floor(width / itemWidth)) : 1;
-  const rowCount = Math.ceil(items.length / colCount);
-
-  const [listProps, currentTabbable] =
-    useVirtualGridArrowNavigation<HTMLDivElement>({
-      colCount,
-      rowCount,
-      itemsCount: sortedItems.length,
-    });
-
-  const onFocusMove = useCallback(
-    (idx) => {
-      const rowIdx = Math.floor(idx / colCount);
-      listRef.current?.scrollToItem(rowIdx);
-    },
-    [listRef, colCount]
-  );
-
-  const rovingFocusProps = useVirtualRovingTabIndex<HTMLDivElement>({
-    currentTabbable,
-    onFocusMove,
-  });
 
   const shouldShowControls = items.length > 0;
+
+  const renderItem: React.ComponentProps<typeof VirtualGrid>['renderItem'] =
+    useCallback(
+      ({ index, ...props }) => {
+        const item = sortedItems[index];
+        return _renderItem({
+          item,
+          viewType,
+          onItemClick,
+          onDeleteItemClick,
+          ...props,
+        });
+      },
+      [_renderItem, onDeleteItemClick, onItemClick, sortedItems, viewType]
+    );
 
   return (
     <ControlsContext.Provider
@@ -280,41 +167,23 @@ export const ItemsGrid = <T extends Item>({
         createControls,
         sortControls: shouldShowControls ? sortControls : null,
         viewTypeControls: shouldShowControls ? viewTypeControls : null,
-        gridProps: mergeProps(
-          { role: 'grid', 'aria-rowcount': rowCount },
-          listProps,
-          rovingFocusProps
-        ),
       }}
     >
-      <div
-        className={container}
+      <VirtualGrid
+        itemMinWidth={itemWidth}
+        itemHeight={itemHeight + spacing[2]}
+        itemsCount={sortedItems.length}
+        colCount={viewType === 'list' ? 1 : undefined}
+        renderItem={renderItem}
+        headerHeight={CONTROLS_HEIGHT}
+        renderHeader={GridControls}
+        classNames={{
+          container,
+          header: controls,
+          row: row,
+        }}
         data-testid={`${itemType}-grid`}
-        {...rectProps}
-      >
-        <FixedSizeList
-          ref={listRef}
-          width={Math.max(width, itemWidth)}
-          height={height}
-          innerElementType={ListWithControls}
-          itemData={{
-            items: sortedItems,
-            viewType,
-            rowsCount: rowCount,
-            columnsCount: colCount,
-            currentTabbable,
-            onItemClick,
-            onDeleteItemClick,
-            renderItem,
-          }}
-          itemCount={rowCount}
-          itemSize={itemHeight + spacing[2]}
-          overscanCount={3}
-        >
-          {/* TypeScript struggles to auto derive generic value without explicitly typing this */}
-          {ItemRow as (props: ItemRowProps<T>) => React.ReactElement}
-        </FixedSizeList>
-      </div>
+      ></VirtualGrid>
     </ControlsContext.Provider>
   );
 };
